refactor: simplify datamart spider to yield AuthorItems
Renamed oa_faculty.py to datamart.py. The spider now just loads faculty
rows from the Datamart and yields one AuthorItem per row with a uw_netid
identifier. All the matching and evidence-saving logic is removed. The
AuthorIdentifierPipeline handles the rest.
